# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restobar Venecia Peruana",
        "description": "Sistema de restaurante en linea. Con control de inventario",
        "contact": {
            "email": "neisserrey@upeu.edu.pe"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://veneciaperuana.dev.com/api/",
            "description": "Documentación del Restobar Venecia Peruana"
        }
    ],
    "paths": {
...
        "/employees": {
            "get": {
                "tags": [
                    "employees"
                ],
                "summary": "Mostrar Empleados",
                "operationId": "App\\Http\\Controllers\\Api\\EmployeeController::index",
                "responses": {
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "200": {
                        "description": "Lista de Empleados. Correcto"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "employees"
                ],
                "summary": "Agregar un nuevo Empleado.",
                "operationId": "store",
                "requestBody": {
                    "description": "Empleado a ser creado",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Employee"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "201": {
                        "description": "Creado"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/employees/{employeeId}": {
            "get": {
                "tags": [
                    "employees"
                ],
                "summary": "Buscar Empleado por ID",
                "description": "Retorna un Empleado",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "employeeId",
                        "in": "path",
                        "description": "ID del Empleado a retornar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Employee"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "404": {
                        "description": "Empleado no existe"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "employees"
                ],
                "summary": "Actualizar un Empleado existente",
                "description": "",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "employeeId",
                        "in": "path",
                        "description": "ID del Empleado a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Empleado a ser actualizada",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Employee"
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Empleado no encontrada"
                    },
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "200": {
                        "description": "Empleado actualizada correctamente"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "employees"
                ],
                "summary": "Elimina un Empleado",
                "description": "",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "employeeId",
                        "in": "path",
                        "description": "Id del Empleado a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Empleado no encontrado"
                    },
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "200": {
                        "description": "Empleado eliminado correctamente"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Employee": {
                "properties": {
                    "name": {
                        "description": "name of the Employee",
                        "type": "string"
                    },
                    "alias": {
                        "description": "alias of the Employee",
                        "type": "string"
                    },
                    "phone": {
                        "description": "phone just 9 numbers",
                        "type": "integer",
                        "format": "int32"
                    },
                    "status": {
                        "description": "status like: created=C, deleted=D",
                        "type": "char"
                    },
                    "created_at": {
                        "description": "date when Employee was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "date when Employee was updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}